requirements.txt
openai==1.56.2
python-dotenv==1.0.1
gradio
ipython